requirements.txt
annotated-types==0.7.0
asttokens==3.0.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
json_repair==0.54.2
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
MouseInfo==0.1.3
nest-asyncio==1.6.0
numpy==2.2.6
opencv-python==4.12.0.88
mss==9.0.1
imageio-ffmpeg==0.4.9
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
pyzmq==27.1.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
PyYAML==6.0.1
